Moved XMLParser to the util folder
