+++ core/src/com/panic/tdt4240/PanicGame.java
@@ -5,7 +5,6 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.panic.tdt4240.models.XMLParser;
 
 public class PanicGame extends ApplicationAdapter {
 	SpriteBatch batch;
